requirements.txt
torch
torchvision
pandas
numpy
pillow
tqdm
scikit-learn
requests